--- coders/composio_coders/__init__.py
@@ -1,1 +1,0 @@
-from .context import Context
