+++ composio_swe/benchmark/run_evaluation.py
@@ -1,8 +1,8 @@
 import logging
 
-from composio_coders.constants import KEY_API_KEY
-from composio_coders.context import Context, set_context
-from composio_coders.swe import CoderAgent, CoderAgentArgs
+from composio_swe.config.constants import KEY_API_KEY
+from composio_swe.cli.context import Context, set_context
+from composio_swe.swe import CoderAgent, CoderAgentArgs
 from datasets import load_dataset
 from rich.logging import RichHandler
