+++ composio_swe/composio_swe/cli/__init__.py
@@ -1,0 +1,1 @@
+from composio_swe.cli.cli import cli
